Add CNAME file to allow github pages deployment
Hi @circa1977, not sure if you're hosting on github pages; we'd like to point a domain at your front end so we can start to share it with folks more easily. Would you be willing to set it up so that the server responds to requests for artbotapp.com? I will take care of adding the DNS record to the zone file. I assume a CNAME pointing the new apex domain to staging.artx.clearbold.com will work? Thanks!